feat: meta statements, maps, and lists
Evaluate mappings and lists. Add support for the "meta" statement which
is how you specify metadata in a Dessert document.

Evaluate identifiers as is. This is intended for mappings which allow
them as keys. In the future, this will likely change to be in mappings
only and identifiers will be evaluated according to the current
namespace instead.