feat: struct statements
Add some work for struct statements. Structs will be the core building
block of a Dessert document and contain a set of fields to represent
each piece of data.

Struct members are no longer stored in a mapping as Go maps are
unordered. Instead, they are stored in an array of key-value structs.

Add some basic types such as unsigned and signed integer types,
float types, and types for reading byte streams, arrays, and matching a
specific string. Subscript access (expr[item]) was added alongside these
types.

Add "endian" and "root" keys to the metadata mapping. These keys provide
the global endianness of the described protocol and the root structure
from which to start reading, respectively.