feat(docs): add specification for 0.1
Add draft for the Dessert language version 0.1. This serves as the
first piece of documentation available for Dessert.

A few things are still pending: many features of the spec reflect
current behavior that will likely change and the current runtime has
features not described in the spec. The spec also uses a notation
language similar to EBNF and RegEx that is currently undocumented.

Also, replace "protocol" with "format" to be more precise on what a
Dessert document describes.